# packages/contracts/openapi/cinewatch-v1.openapi.json
{
  "components": {
    "schemas": {
      "CatalogBrowseResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogMediaSummary"
            },
            "maxItems": 24,
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "enum": [
              "genre",
              "collection",
              "country"
            ],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "media_type": {
            "enum": [
              "all",
              "movie",
              "tv"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "slug": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          },
          "sort": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Sort",
            "type": "string"
          },
          "title": {
            "maxLength": 220,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          },
          "total_results": {
            "minimum": 0.0,
            "title": "Total Results",
            "type": "integer"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 2200.0,
                "minimum": 1800.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "kind",
          "slug",
          "title",
          "page",
          "total_pages",
          "total_results",
          "media_type",
          "sort"
        ],
        "title": "CatalogBrowseResponse",
        "type": "object"
      },
...
      "CatalogPersonResponse": {
        "additionalProperties": false,
        "properties": {
          "biography": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Biography"
          },
          "birthday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/CatalogPersonCredit"
            },
            "maxItems": 40,
            "title": "Credits",
            "type": "array"
          },
          "crew_credits": {
            "items": {
              "$ref": "#/components/schemas/CatalogPersonCredit"
            },
            "maxItems": 40,
            "title": "Crew Credits",
            "type": "array"
          },
          "deathday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deathday"
          },
          "hero_backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Backdrop Url"
          },
          "known_for_department": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known For Department"
          },
          "name": {
            "maxLength": 220,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "place_of_birth": {
            "anyOf": [
              {
                "maxLength": 260,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Of Birth"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "provider_id": {
            "exclusiveMinimum": 0.0,
            "title": "Provider Id",
            "type": "integer"
          }
        },
        "required": [
          "provider_id",
          "name"
        ],
        "title": "CatalogPersonResponse",
        "type": "object"
      },
      "CatalogProviderResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogMediaSummary"
            },
            "maxItems": 24,
            "title": "Items",
            "type": "array"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "name": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "official_homepage_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Official Homepage Url"
          },
          "provider_id": {
            "exclusiveMinimum": 0.0,
            "title": "Provider Id",
            "type": "integer"
          },
          "region": {
            "maxLength": 2,
            "minLength": 2,
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "provider_id",
          "name",
          "region"
        ],
        "title": "CatalogProviderResponse",
        "type": "object"
      },
...
      "CatalogSeasonResponse": {
        "additionalProperties": false,
        "properties": {
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/CatalogEpisode"
            },
            "maxItems": 100,
            "title": "Episodes",
            "type": "array"
          },
          "name": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "provider_id": {
            "exclusiveMinimum": 0.0,
            "title": "Provider Id",
            "type": "integer"
          },
          "season_number": {
            "minimum": 0.0,
            "title": "Season Number",
            "type": "integer"
          }
        },
        "required": [
          "provider_id",
          "season_number",
          "name"
        ],
        "title": "CatalogSeasonResponse",
        "type": "object"
      },
      "CatalogTitleResponse": {
        "additionalProperties": false,
        "properties": {
          "awards": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards"
          },
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "box_office": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Box Office"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CatalogCastMember"
            },
            "maxItems": 18,
            "title": "Cast",
            "type": "array"
          },
          "creators": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Creators",
            "type": "array"
          },
          "crew": {
            "items": {
              "$ref": "#/components/schemas/CatalogCrewMember"
            },
            "maxItems": 30,
            "title": "Crew",
            "type": "array"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "media_type": {
            "enum": [
              "movie",
              "tv"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "networks": {
            "items": {
              "$ref": "#/components/schemas/CatalogNetwork"
            },
            "maxItems": 12,
            "title": "Networks",
            "type": "array"
          },
          "original_language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Language"
          },
          "original_title": {
            "anyOf": [
              {
                "maxLength": 260,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "production_budget": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Budget"
          },
          "provider_id": {
            "exclusiveMinimum": 0.0,
            "title": "Provider Id",
            "type": "integer"
          },
          "ratings": {
            "items": {
              "$ref": "#/components/schemas/CatalogRating"
            },
            "maxItems": 4,
            "title": "Ratings",
            "type": "array"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/CatalogMediaSummary"
            },
            "maxItems": 12,
            "title": "Recommendations",
            "type": "array"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "revenue": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue"
          },
          "review_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Review Count",
            "type": "integer"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/CatalogReview"
            },
            "maxItems": 3,
            "title": "Reviews",
            "type": "array"
          },
          "runtime_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Minutes"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/CatalogSeason"
            },
            "maxItems": 120,
            "title": "Seasons",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tagline": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "title": {
            "maxLength": 260,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "trailer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogTrailer"
              },
              {
                "type": "null"
              }
            ]
          },
          "watch_information_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Information Url"
          },
          "watch_providers": {
            "items": {
              "$ref": "#/components/schemas/CatalogWatchProvider"
            },
            "maxItems": 18,
            "title": "Watch Providers",
            "type": "array"
          },
          "watch_region": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Region"
          },
          "writers": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Writers",
            "type": "array"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 2200.0,
                "minimum": 1800.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "provider_id",
          "media_type",
          "title"
        ],
        "title": "CatalogTitleResponse",
        "type": "object"
      },
...
      "CatalogVideosResponse": {
        "additionalProperties": false,
        "properties": {
          "media_type": {
            "enum": [
              "movie",
              "tv"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "provider_id": {
            "exclusiveMinimum": 0.0,
            "title": "Provider Id",
            "type": "integer"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/CatalogTrailer"
            },
            "maxItems": 20,
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "provider_id",
          "media_type"
        ],
        "title": "CatalogVideosResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "CineWatch TV V1 backend service boundary.",
    "title": "CineWatch TV API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/catalog/browse/{kind}/{slug}": {
      "get": {
        "operationId": "v1_catalog_browse",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "genre",
                "collection",
                "country"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "movie",
                "tv"
              ],
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2200,
                  "minimum": 1800,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 12,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "popularity.desc",
              "pattern": "^(popularity|vote_average|primary_release_date|first_air_date)\\.(asc|desc)$",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogBrowseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Filtered CineWatch discovery browse",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/person/{provider_id}": {
      "get": {
        "operationId": "v1_catalog_person",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch person details",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/provider/{provider_id}": {
      "get": {
        "operationId": "v1_catalog_provider",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "US",
              "maxLength": 2,
              "minLength": 2,
              "title": "Region",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch streaming provider discovery",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/title/tv/{provider_id}/season/{season_number}": {
      "get": {
        "operationId": "v1_catalog_tv_season",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "season_number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Season Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogSeasonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch television season episodes",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/title/{media_type}/{provider_id}": {
      "get": {
        "operationId": "v1_catalog_title",
        "parameters": [
          {
            "in": "path",
            "name": "media_type",
            "required": true,
            "schema": {
              "enum": [
                "movie",
                "tv"
              ],
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogTitleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch title details",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/title/{media_type}/{provider_id}/reviews": {
      "get": {
        "operationId": "v1_catalog_title_reviews",
        "parameters": [
          {
            "in": "path",
            "name": "media_type",
            "required": true,
            "schema": {
              "enum": [
                "movie",
                "tv"
              ],
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch title reviews",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/title/{media_type}/{provider_id}/videos": {
      "get": {
        "operationId": "v1_catalog_title_videos",
        "parameters": [
          {
            "in": "path",
            "name": "media_type",
            "required": true,
            "schema": {
              "enum": [
                "movie",
                "tv"
              ],
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Provider Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogVideosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "CineWatch title videos",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}